chore(daemon): rename echolalia to elsi
