+++ daemon/pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "analysis-service-daemon"
 version = "0.2.0"
-description = "analysis service daemon for Echolalia"
+description = "analysis service daemon for ELSI"
 authors = [
     {name = "Lawrence Borst",email = "laurens.borst@ens.psl.eu"}
 ]
